Module Caqti_async

Connector for Async.

include Caqti_connect_sig.S with type 'a future := 'a Async_kernel.Deferred.t

type +'a future

The type of a deferred value of type 'a.

module Pool : Caqti_pool_sig.S with type 'a future := 'a future

A pool implementation for the current concurrency library.

module Stream : Caqti_stream.S with type 'a future := 'a future

module type CONNECTION = Caqti_connection_sig.S with type 'a future := 'a future and type ('a, 'err) stream := ('a, 'err) Stream.t

The connection API specialized for the current concurrency library.

type connection = (module CONNECTION)

Shortcut for the connection API passed as a value.

val connect : Uri.t -> (connection, [> Caqti_error.load_or_connect ]) Stdlib.result future

connect uri locates and loads a driver which can handle uri, passes uri to the driver, which establish a connection and returns a first-class module implementing Caqti_connection_sig.S.

val connect_pool : ?⁠max_size:int -> Uri.t -> ((connection, [> Caqti_error.connect ]) Pool.t, [> Caqti_error.load ]) Stdlib.result

connect_pool uri is a pool of database connections constructed by connect uri.