Module Caqti_async

Connector for Async.

module Stream : Caqti_stream_sig.S with type 'a fiber := 'a Async_kernel.Deferred.t

module Pool : Caqti_pool_sig.S with type 'a fiber := 'a Async_kernel.Deferred.t

module type CONNECTION = Caqti_connection_sig.S with type 'a fiber := 'a Async_kernel.Deferred.t and type ('a, 'e) stream := ('a, 'e) Stream.t

include Caqti_connect_sig.S with type 'a fiber := 'a Async_kernel.Deferred.t and type 'a with_switch := 'a and type 'a with_stdenv := 'a and type ('a, 'e) stream := ('a, 'e) Stream.t and type ('a, 'e) pool := ('a, 'e) Pool.t and type connection = (module CONNECTION)

type connection = (module CONNECTION)

Shortcut for the connection module when passed as a value.

val connect : ?env:(Caqti_driver_info.t -> string -> Caqti_query.t) -> ?config:Caqti_connect_config.t -> ?tweaks_version:(int * int) -> Uri.t -> (connection, [> Caqti_error.load_or_connect ]) Stdlib.result Async_kernel.Deferred.t

connect uri locates and loads a driver which can handle uri, passes uri to the driver, which establish a connection and returns a first-class module implementing Caqti_connection_sig.S.

connect uri connects to the database at uri and returns a first class module implementing Caqti_connection_sig.S for the given database system. In case of preemptive threading, the connection must only be used from the thread where it was created.

The correct driver for the database system is inferred from the schema of uri; see the respective drivers for the supported schemas and related URI syntax. A driver can either be linked in to the application or, if supported, dynamically linked using the caqti-dynload package.

parameter env
If provided, this function will do a final expansion of environment variables which occurs in the query templates of the requests executed on the connection.

parameter config
Configuration parameters related to the interaction with the database.

parameter tweaks_version

deprecated
This should now be passed via the config parameter using the Caqti_connect_config.tweaks_version key.

val with_connection : ?env:(Caqti_driver_info.t -> string -> Caqti_query.t) -> ?config:Caqti_connect_config.t -> ?tweaks_version:(int * int) -> Uri.t -> (connection -> ('a, [> Caqti_error.load_or_connect ] as 'e) Stdlib.result Async_kernel.Deferred.t) -> ('a, 'e) Stdlib.result Async_kernel.Deferred.t

with_connection uri f calls connect on uri. If connect evaluates to Ok connection, with_connection passes the connection to f. Once f either evaluates to a result, or raises an exception, with_connection closes the database connection.

parameter env
Passed to connect.

parameter config
Passed to connect.

parameter tweaks_version

deprecated
This should now be passed via the config parameter using the Caqti_connect_config.tweaks_version key.

val connect_pool : ?pool_config:Caqti_pool_config.t -> ?post_connect: (connection -> (unit, 'connect_error) Stdlib.result Async_kernel.Deferred.t) -> ?env:(Caqti_driver_info.t -> string -> Caqti_query.t) -> ?config:Caqti_connect_config.t -> ?tweaks_version:(int * int) -> Uri.t -> ((connection, [> Caqti_error.connect ] as 'connect_error) Pool.t, [> Caqti_error.load ]) Stdlib.result

connect_pool uri is a pool of database connections constructed by connect uri.

Do not use pooling for connections to volatile resources like sqlite3::memory: and beware of temporary tables or other objects which may not be shared across connections to the same URI.

If you use preemptive threading, note that the connection pool must only be used from the thread where it was created. Use thread local storage to create a separate pool per thread if necessary.

parameter pool_config
Provides tuning parameters for the pool. The default is the result of a fresh call of Caqti_pool_config.default_from_env.

parameter post_connect
A task to run after establishing a new connection and before the connection becomes available to the application. This function can be used to customize to the database session.

parameter config
Passed to connect when creating new connections.

parameter env
Passed to connect when creating new connections.

parameter tweaks_version

deprecated
This should now be passed via the config parameter using the Caqti_connect_config.tweaks_version key.